:art: Move callback class to it`s package
