+++ app/src/main/java/ufrj/devmob/votadevmob/poll/PollModel.kt
@@ -1,7 +1,7 @@
 package ufrj.devmob.votadevmob.poll
 
 import com.google.firebase.firestore.FirebaseFirestore
-import ufrj.devmob.votadevmob.Callback
+import ufrj.devmob.votadevmob.core.callback.Callback
 import ufrj.devmob.votadevmob.core.model.Vote
 
 class PollModel {
